03-refresh_not_persisted/pyproject.toml
[project]
name = "03-refresh-not-persisted"
version = "0.1.0"
description = "ADK bug reproduction: ToolAuthHandler._get_existing_credential refreshes OAuth2 credentials in memory but does not persist them"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi>=0.135.3,<0.136.0",
    "google-adk==1.28.0",
    "httpx>=0.28.1,<0.29.0",
    "python-dotenv>=1.2.2,<2.0.0",
    "uvicorn>=0.44.0,<0.45.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.10,<0.16.0",
]

[tool.ruff]
target-version = "py313"
line-length = 88
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "S",   # flake8-bandit
    "SIM", # flake8-simplify
    "PTH", # flake8-use-pathlib
    "RET", # flake8-return
]
